+++ test/legacy_test/test_model.py
@@ -244,7 +244,8 @@ class TestModel(unittest.TestCase):
         self.fit_with_tuple_input(True)
 
     def test_fit_static_with_tuple_input(self):
-        self.fit_with_tuple_input(False)
+        if not paddle.framework.in_pir_mode():
+            self.fit_with_tuple_input(False)
 
     def test_fit_dynamic_with_rank(self):
         self.fit(True, 2, 0)
@@ -268,7 +269,8 @@ class TestModel(unittest.TestCase):
         self.predict(True)
 
     def test_predict_static(self):
-        self.predict(False)
+        with paddle.pir_utils.OldIrGuard():
+            self.predict(False)
 
     def test_prepare_context(self):
         prepare_distributed_context()
@@ -279,64 +281,65 @@ class TestModel(unittest.TestCase):
         paddle.seed(seed)
         paddle.framework.random._manual_program_seed(seed)
 
-        net = LeNet()
-        optim_new = paddle.optimizer.Adam(
-            learning_rate=0.001, parameters=net.parameters()
-        )
-        model = Model(net, inputs=self.inputs, labels=self.labels)
-        model.prepare(
-            optim_new,
-            loss=CrossEntropyLoss(reduction="sum"),
-            metrics=Accuracy(),
-        )
-        model.fit(self.train_dataset, batch_size=64, shuffle=False)
+        with paddle.pir_utils.OldIrGuard():
+            net = LeNet()
+            optim_new = paddle.optimizer.Adam(
+                learning_rate=0.001, parameters=net.parameters()
+            )
+            model = Model(net, inputs=self.inputs, labels=self.labels)
+            model.prepare(
+                optim_new,
+                loss=CrossEntropyLoss(reduction="sum"),
+                metrics=Accuracy(),
+            )
+            model.fit(self.train_dataset, batch_size=64, shuffle=False)
 
-        result = model.evaluate(self.val_dataset, batch_size=64)
-        np.testing.assert_allclose(result['acc'], self.acc1)
+            result = model.evaluate(self.val_dataset, batch_size=64)
+            np.testing.assert_allclose(result['acc'], self.acc1)
 
-        model.fit(
-            self.train_dataset,
-            batch_size=64,
-            shuffle=False,
-            num_iters=num_iters,
-        )
+            model.fit(
+                self.train_dataset,
+                batch_size=64,
+                shuffle=False,
+                num_iters=num_iters,
+            )
 
-        result = model.evaluate(
-            self.val_dataset, batch_size=64, num_iters=num_iters
-        )
+            result = model.evaluate(
+                self.val_dataset, batch_size=64, num_iters=num_iters
+            )
 
-        model.fit(self.train_dataset, batch_size=(64, 64), shuffle=False)
+            model.fit(self.train_dataset, batch_size=(64, 64), shuffle=False)
 
-        train_sampler = DistributedBatchSampler(
-            self.train_dataset,
-            batch_size=64,
-            shuffle=False,
-            num_replicas=num_replicas,
-            rank=rank,
-        )
-        val_sampler = DistributedBatchSampler(
-            self.val_dataset,
-            batch_size=64,
-            shuffle=False,
-            num_replicas=num_replicas,
-            rank=rank,
-        )
+            train_sampler = DistributedBatchSampler(
+                self.train_dataset,
+                batch_size=64,
+                shuffle=False,
+                num_replicas=num_replicas,
+                rank=rank,
+            )
+            val_sampler = DistributedBatchSampler(
+                self.val_dataset,
+                batch_size=64,
+                shuffle=False,
+                num_replicas=num_replicas,
+                rank=rank,
+            )
 
-        train_loader = paddle.io.DataLoader(
-            self.train_dataset,
-            batch_sampler=train_sampler,
-            places=self.device,
-            return_list=True,
-        )
+            train_loader = paddle.io.DataLoader(
+                self.train_dataset,
+                batch_sampler=train_sampler,
+                places=self.device,
+                return_list=True,
+            )
 
-        val_loader = paddle.io.DataLoader(
-            self.val_dataset,
-            batch_sampler=val_sampler,
-            places=self.device,
-            return_list=True,
-        )
+            val_loader = paddle.io.DataLoader(
+                self.val_dataset,
+                batch_sampler=val_sampler,
+                places=self.device,
+                return_list=True,
+            )
 
-        model.fit(train_loader, val_loader)
+            model.fit(train_loader, val_loader)
         base.disable_dygraph() if dynamic else None
 
     def fit_with_tuple_input(self, dynamic, num_replicas=None, rank=None):
@@ -394,26 +397,27 @@ class TestModel(unittest.TestCase):
 
     def evaluate(self, dynamic):
         base.enable_dygraph(self.device) if dynamic else None
-        model = Model(LeNet(), self.inputs, self.labels)
-        model.prepare(metrics=Accuracy())
-        model.load(self.weight_path)
-        result = model.evaluate(self.val_dataset, batch_size=64)
-        np.testing.assert_allclose(result['acc'], self.acc1)
+        with paddle.pir_utils.OldIrGuard():
+            model = Model(LeNet(), self.inputs, self.labels)
+            model.prepare(metrics=Accuracy())
+            model.load(self.weight_path)
+            result = model.evaluate(self.val_dataset, batch_size=64)
+            np.testing.assert_allclose(result['acc'], self.acc1)
 
-        sampler = DistributedBatchSampler(
-            self.val_dataset, batch_size=64, shuffle=False
-        )
+            sampler = DistributedBatchSampler(
+                self.val_dataset, batch_size=64, shuffle=False
+            )
 
-        val_loader = paddle.io.DataLoader(
-            self.val_dataset,
-            batch_sampler=sampler,
-            places=self.device,
-            return_list=True,
-        )
+            val_loader = paddle.io.DataLoader(
+                self.val_dataset,
+                batch_sampler=sampler,
+                places=self.device,
+                return_list=True,
+            )
 
-        model.evaluate(val_loader)
+            model.evaluate(val_loader)
 
-        base.disable_dygraph() if dynamic else None
+            base.disable_dygraph() if dynamic else None
 
     def predict(self, dynamic):
         base.enable_dygraph(self.device) if dynamic else None
@@ -516,18 +520,19 @@ class TestModelFunction(unittest.TestCase):
             base.enable_dygraph(device) if dynamic else None
             self.set_seed()
 
-            net = MyModel()
-            optim2 = paddle.optimizer.SGD(
-                learning_rate=0.001, parameters=net.parameters()
-            )
+            with paddle.pir_utils.OldIrGuard():
+                net = MyModel()
+                optim2 = paddle.optimizer.SGD(
+                    learning_rate=0.001, parameters=net.parameters()
+                )
 
-            inputs = [InputSpec([None, dim], 'float32', 'x')]
-            labels = [InputSpec([None, 1], 'int64', 'label')]
-            model = Model(net, inputs, labels)
-            model.prepare(optim2, loss=CrossEntropyLoss(reduction="sum"))
-            (loss,) = model.train_batch([data], [label])
-            np.testing.assert_allclose(loss.flatten(), ref.flatten())
-            base.disable_dygraph() if dynamic else None
+                inputs = [InputSpec([None, dim], 'float32', 'x')]
+                labels = [InputSpec([None, 1], 'int64', 'label')]
+                model = Model(net, inputs, labels)
+                model.prepare(optim2, loss=CrossEntropyLoss(reduction="sum"))
+                (loss,) = model.train_batch([data], [label])
+                np.testing.assert_allclose(loss.flatten(), ref.flatten())
+                base.disable_dygraph() if dynamic else None
 
     def test_test_batch(self):
         dim = 20
@@ -547,14 +552,16 @@ class TestModelFunction(unittest.TestCase):
             device = paddle.set_device('cpu')
             base.enable_dygraph(device) if dynamic else None
             self.set_seed()
-            net = MyModel()
-            inputs = [InputSpec([None, dim], 'float32', 'x')]
-            model = Model(net, inputs)
-            model.prepare()
-            (out,) = model.predict_batch([data])
 
-            np.testing.assert_allclose(out, ref, rtol=1e-6)
-            base.disable_dygraph() if dynamic else None
+            with paddle.pir_utils.OldIrGuard():
+                net = MyModel()
+                inputs = [InputSpec([None, dim], 'float32', 'x')]
+                model = Model(net, inputs)
+                model.prepare()
+                (out,) = model.predict_batch([data])
+
+                np.testing.assert_allclose(out, ref, rtol=1e-6)
+                base.disable_dygraph() if dynamic else None
 
     def test_save_load(self):
         path = os.path.join(tempfile.mkdtemp(), '.cache_test_save_load')
@@ -563,19 +570,21 @@ class TestModelFunction(unittest.TestCase):
         for dynamic in [True, False]:
             device = paddle.set_device('cpu')
             base.enable_dygraph(device) if dynamic else None
-            net = MyModel()
-            inputs = [InputSpec([None, 20], 'float32', 'x')]
-            labels = [InputSpec([None, 1], 'int64', 'label')]
-            optim = paddle.optimizer.SGD(
-                learning_rate=0.001, parameters=net.parameters()
-            )
-            model = Model(net, inputs, labels)
-            model.prepare(
-                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
-            )
-            model.save(path)
-            model.load(path)
-            base.disable_dygraph() if dynamic else None
+
+            with paddle.pir_utils.OldIrGuard():
+                net = MyModel()
+                inputs = [InputSpec([None, 20], 'float32', 'x')]
+                labels = [InputSpec([None, 1], 'int64', 'label')]
+                optim = paddle.optimizer.SGD(
+                    learning_rate=0.001, parameters=net.parameters()
+                )
+                model = Model(net, inputs, labels)
+                model.prepare(
+                    optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+                )
+                model.save(path)
+                model.load(path)
+                base.disable_dygraph() if dynamic else None
         shutil.rmtree(path)
 
     def test_dynamic_load(self):
@@ -620,15 +629,18 @@ class TestModelFunction(unittest.TestCase):
         model.save(path)
         base.disable_dygraph()
 
-        inputs = [InputSpec([None, 20], 'float32', 'x')]
-        labels = [InputSpec([None, 1], 'int64', 'label')]
-        model = Model(MyModel(), inputs, labels)
-        optim = paddle.optimizer.SGD(
-            learning_rate=0.001, parameters=model.parameters()
-        )
-        model.prepare(optimizer=optim, loss=CrossEntropyLoss(reduction="sum"))
-        model.load(path)
-        shutil.rmtree(path)
+        with paddle.pir_utils.OldIrGuard():
+            inputs = [InputSpec([None, 20], 'float32', 'x')]
+            labels = [InputSpec([None, 1], 'int64', 'label')]
+            model = Model(MyModel(), inputs, labels)
+            optim = paddle.optimizer.SGD(
+                learning_rate=0.001, parameters=model.parameters()
+            )
+            model.prepare(
+                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+            )
+            model.load(path)
+            shutil.rmtree(path)
 
     def test_static_save_dynamic_load(self):
         path = os.path.join(
@@ -636,43 +648,51 @@ class TestModelFunction(unittest.TestCase):
         )
         if not os.path.exists(path):
             os.makedirs(path)
-        net = MyModel()
-        inputs = [InputSpec([None, 20], 'float32', 'x')]
-        labels = [InputSpec([None, 1], 'int64', 'label')]
-        optim = paddle.optimizer.SGD(
-            learning_rate=0.001, parameters=net.parameters()
-        )
-        model = Model(net, inputs, labels)
-        model.prepare(optimizer=optim, loss=CrossEntropyLoss(reduction="sum"))
-        model.save(path)
 
-        device = paddle.set_device('cpu')
-        base.enable_dygraph(device)  # if dynamic else None
+        with paddle.pir_utils.OldIrGuard():
+            net = MyModel()
+            inputs = [InputSpec([None, 20], 'float32', 'x')]
+            labels = [InputSpec([None, 1], 'int64', 'label')]
+            optim = paddle.optimizer.SGD(
+                learning_rate=0.001, parameters=net.parameters()
+            )
+            model = Model(net, inputs, labels)
+            model.prepare(
+                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+            )
+            model.save(path)
 
-        net = MyModel()
-        inputs = [InputSpec([None, 20], 'float32', 'x')]
-        labels = [InputSpec([None, 1], 'int64', 'label')]
-        optim = paddle.optimizer.SGD(
-            learning_rate=0.001, parameters=net.parameters()
-        )
-        model = Model(net, inputs, labels)
-        model.prepare(optimizer=optim, loss=CrossEntropyLoss(reduction="sum"))
-        model.load(path)
-        shutil.rmtree(path)
-        base.disable_dygraph()
+            device = paddle.set_device('cpu')
+            base.enable_dygraph(device)  # if dynamic else None
+
+            net = MyModel()
+            inputs = [InputSpec([None, 20], 'float32', 'x')]
+            labels = [InputSpec([None, 1], 'int64', 'label')]
+            optim = paddle.optimizer.SGD(
+                learning_rate=0.001, parameters=net.parameters()
+            )
+            model = Model(net, inputs, labels)
+            model.prepare(
+                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+            )
+            model.load(path)
+            shutil.rmtree(path)
+            base.disable_dygraph()
 
     def test_parameters(self):
         for dynamic in [True, False]:
             device = paddle.set_device('cpu')
             base.enable_dygraph(device) if dynamic else None
-            net = MyModel()
-            inputs = [InputSpec([None, 20], 'float32', 'x')]
-            model = Model(net, inputs)
-            model.prepare()
-            params = model.parameters()
-            self.assertTrue(params[0].shape[0] == 20)
-            self.assertTrue(params[0].shape[1] == 10)
-            base.disable_dygraph() if dynamic else None
+
+            with paddle.pir_utils.OldIrGuard():
+                net = MyModel()
+                inputs = [InputSpec([None, 20], 'float32', 'x')]
+                model = Model(net, inputs)
+                model.prepare()
+                params = model.parameters()
+                self.assertTrue(params[0].shape[0] == 20)
+                self.assertTrue(params[0].shape[1] == 10)
+                base.disable_dygraph() if dynamic else None
 
     def test_summary(self):
         def _get_param_from_state_dict(state_dict):
@@ -820,42 +840,43 @@ class TestModelFunction(unittest.TestCase):
         for dynamic in [True, False]:
             paddle.disable_static() if dynamic else None
 
-            net = LeNet()
-            inputs = [InputSpec([None, 1, 28, 28], 'float32', 'x')]
-            model = Model(net, inputs)
-            model.prepare()
+            if not paddle.framework.in_pir_mode():
+                net = LeNet()
+                inputs = [InputSpec([None, 1, 28, 28], 'float32', 'x')]
+                model = Model(net, inputs)
+                model.prepare()
 
-            tensor_img = np.array(
-                np.random.random((1, 1, 28, 28)), dtype=np.float32
-            )
+                tensor_img = np.array(
+                    np.random.random((1, 1, 28, 28)), dtype=np.float32
+                )
 
-            model.save(save_dir, training=False)
-            ori_results = model.predict_batch(tensor_img)
-            base.disable_dygraph() if dynamic else None
+                model.save(save_dir, training=False)
+                ori_results = model.predict_batch(tensor_img)
+                base.disable_dygraph() if dynamic else None
 
-            place = (
-                base.CPUPlace()
-                if not base.is_compiled_with_cuda()
-                else base.CUDAPlace(0)
-            )
-            new_scope = base.Scope()
-            with base.scope_guard(new_scope):
-                exe = base.Executor(place)
-                [
-                    inference_program,
-                    feed_target_names,
-                    fetch_targets,
-                ] = paddle.static.io.load_inference_model(
-                    path_prefix=save_dir, executor=exe
+                place = (
+                    base.CPUPlace()
+                    if not base.is_compiled_with_cuda()
+                    else base.CUDAPlace(0)
                 )
-                results = exe.run(
-                    inference_program,
-                    feed={feed_target_names[0]: tensor_img},
-                    fetch_list=fetch_targets,
-                )
-                np.testing.assert_allclose(
-                    results, ori_results, rtol=1e-5, atol=1e-6
-                )
+                new_scope = base.Scope()
+                with base.scope_guard(new_scope):
+                    exe = base.Executor(place)
+                    [
+                        inference_program,
+                        feed_target_names,
+                        fetch_targets,
+                    ] = paddle.static.io.load_inference_model(
+                        path_prefix=save_dir, executor=exe
+                    )
+                    results = exe.run(
+                        inference_program,
+                        feed={feed_target_names[0]: tensor_img},
+                        fetch_list=fetch_targets,
+                    )
+                    np.testing.assert_allclose(
+                        results, ori_results, rtol=1e-5, atol=1e-6
+                    )
 
             paddle.enable_static()
 
@@ -995,21 +1016,23 @@ class TestModelWithLRScheduler(unittest.TestCase):
         )
         # static test
         paddle.enable_static()
+        with paddle.pir_utils.OldIrGuard():
+            net = MyModel()
+            inputs = [InputSpec([None, 20], 'float32', 'x')]
+            labels = [InputSpec([None, 1], 'int64', 'label')]
+            optim = make_optimizer(net.parameters())
+            model = Model(net, inputs, labels)
+            model.prepare(
+                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+            )
 
-        net = MyModel()
-        inputs = [InputSpec([None, 20], 'float32', 'x')]
-        labels = [InputSpec([None, 1], 'int64', 'label')]
-        optim = make_optimizer(net.parameters())
-        model = Model(net, inputs, labels)
-        model.prepare(optimizer=optim, loss=CrossEntropyLoss(reduction="sum"))
+            dataset = MyDataset()
+            model.fit(dataset, dataset, batch_size=4, epochs=10, num_workers=0)
 
-        dataset = MyDataset()
-        model.fit(dataset, dataset, batch_size=4, epochs=10, num_workers=0)
-
-        np.testing.assert_allclose(
-            model._optimizer._learning_rate.last_lr,
-            base_lr * (0.1 ** len(boundaries)),
-        )
+            np.testing.assert_allclose(
+                model._optimizer._learning_rate.last_lr,
+                base_lr * (0.1 ** len(boundaries)),
+            )
 
     def test_fit_by_epoch(self):
         base_lr = 1e-3
@@ -1075,36 +1098,39 @@ class TestModelWithLRScheduler(unittest.TestCase):
         # static test
         paddle.enable_static()
 
-        net = MyModel()
-        inputs = [InputSpec([None, 20], 'float32', 'x')]
-        labels = [InputSpec([None, 1], 'int64', 'label')]
-        optim = make_optimizer(net.parameters())
-        model = Model(net, inputs, labels)
-        model.prepare(optimizer=optim, loss=CrossEntropyLoss(reduction="sum"))
+        with paddle.pir_utils.OldIrGuard():
+            net = MyModel()
+            inputs = [InputSpec([None, 20], 'float32', 'x')]
+            labels = [InputSpec([None, 1], 'int64', 'label')]
+            optim = make_optimizer(net.parameters())
+            model = Model(net, inputs, labels)
+            model.prepare(
+                optimizer=optim, loss=CrossEntropyLoss(reduction="sum")
+            )
 
-        dataset = MyDataset()
+            dataset = MyDataset()
 
-        lr_scheduler_callback = paddle.callbacks.LRScheduler(
-            by_step=False, by_epoch=True
-        )
+            lr_scheduler_callback = paddle.callbacks.LRScheduler(
+                by_step=False, by_epoch=True
+            )
 
-        model.fit(
-            dataset,
-            dataset,
-            batch_size=4,
-            epochs=epochs,
-            num_workers=0,
-            callbacks=lr_scheduler_callback,
-        )
+            model.fit(
+                dataset,
+                dataset,
+                batch_size=4,
+                epochs=epochs,
+                num_workers=0,
+                callbacks=lr_scheduler_callback,
+            )
 
-        cnt = 0
-        for b in boundaries:
-            if b + warmup_epochs <= epochs:
-                cnt += 1
+            cnt = 0
+            for b in boundaries:
+                if b + warmup_epochs <= epochs:
+                    cnt += 1
 
-        np.testing.assert_allclose(
-            model._optimizer._learning_rate.last_lr, base_lr * (0.1**cnt)
-        )
+            np.testing.assert_allclose(
+                model._optimizer._learning_rate.last_lr, base_lr * (0.1**cnt)
+            )
 
 
 class TestRaiseError(unittest.TestCase):
@@ -1135,16 +1161,17 @@ class TestRaiseError(unittest.TestCase):
 
     def test_save_infer_model_without_file_prefix(self):
         paddle.enable_static()
-        net = LeNet()
-        inputs = [InputSpec([None, 1, 28, 28], 'float32', 'x')]
-        model = Model(net, inputs)
-        model.prepare()
-        path = ""
-        tensor_img = np.array(
-            np.random.random((1, 1, 28, 28)), dtype=np.float32
-        )
-        with self.assertRaises(ValueError):
-            model.save(path, training=False)
+        with paddle.pir_utils.OldIrGuard():
+            net = LeNet()
+            inputs = [InputSpec([None, 1, 28, 28], 'float32', 'x')]
+            model = Model(net, inputs)
+            model.prepare()
+            path = ""
+            tensor_img = np.array(
+                np.random.random((1, 1, 28, 28)), dtype=np.float32
+            )
+            with self.assertRaises(ValueError):
+                model.save(path, training=False)
 
 
 if __name__ == '__main__':
